+++ doc/versionHistory.rst
@@ -27,6 +27,14 @@ Version History
 1.6.8
 -------------
 
+* When aggregating donut stamps, only keep the visit level metadata.
+
+.._lsst.ts.donut.viz-1.6.8
+
+-------------
+1.6.8
+-------------
+
 * Switch to post_isr_image and (in some pipelines) v2 steps
 
 .._lsst.ts.donut.viz-1.6.7